backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
geopy==2.4.1
numpy==1.24.4
pymysql==1.0.3
